+++ tests/Loader/PackageTest.php
@@ -24,7 +24,7 @@ class PackageTest extends TestCase
 
         $this->assertInstanceOf(Model::class, $package);
         $this->assertSame('innmind/url', $package->name()->toString());
-        $this->assertSame('4.1.0', $package->version()->toString());
+        $this->assertSame('4.3.0', $package->version()->toString());
         $this->assertSame(
             'https://packagist.org/packages/innmind/url',
             $package->packagist()->toString(),
@@ -42,7 +42,7 @@ class PackageTest extends TestCase
         );
 
         $this->assertInstanceOf(Model::class, $package);
-        $this->assertSame('7.5.0', $package->version()->toString());
+        $this->assertSame('7.8.0', $package->version()->toString());
     }
 
     public function testReturnNothingWhenThePackageDoesNotExist()